+++ src/core/controllers/WebSocketController.ts
@@ -8,27 +8,35 @@ export type WebSocketNextHandler = (controller: WebSocketController) => void;
 /**
  * WebSocketController manages messages come from a socket.io client.
  * 
- * When you define a method in a WebSocketController, it will be called 
- * automatically when a socket.io event fires. All methods in a 
- * WebSocketController accept two or more parameters in the following sequence: 
+ * When you define a method in a WebSocketController and bind it to a certain 
+ * socket.io event, it will be called automatically when the event fires. All 
+ * methods in a WebSocketController accept at least one parameter: 
  * 
  * - `...data: any[]` sent by the client.
- * - `socket: WebSocket` the underlying socket.
  * 
- * If you're programing with pure JavaScript, I suggest you call `socket` from
- * the `this` object in the controller instead, it should give you proper IDE 
- * hints of this two objects.
+ * The corresponding `socket: WebSocket` object can be auto-injected into the 
+ * controller as well, and its position is arbitrary. Or you can call it from 
+ * `this`:
  * 
- * The decorator function `@event()` is used to set socket events,but when 
- * you're using pure JavaScript, there is not decorators. but the framework 
+ * `let { socket } = this;`
+ * 
+ * You may `return` some data inside a method that bound to an event, when the
+ * method is called, they will be automatically sent to the client. Actions 
+ * will be handled in a Promise constructor, so you can do what ever you want 
+ * inside the method. Using `async` methods to do asynchronous operations is 
+ * recommended.
+ * 
+ * If you want to send a response manually, you can just call the 
+ * `socket.emit()` to do so, the WebSocket support continuous message 
+ * transmissions.
+ * 
+ * The decorator function `@event()` is used to set socket events. But when 
+ * you're coding in JavaScript, there is not decorators, the framework 
  * support another compatible way to allow you doing such things by using the 
  * **jsdoc** block with an `@event` tag, but you need to set 
  * `config.enableDocRoute` to `true`.
  */
 export class WebSocketController extends Controller {
-    /** A reference to the class object. */
-    Class: typeof WebSocketController = <any>this.constructor;
-
     /** Reference to the corresponding socket context. */
     readonly socket: WebSocket;
     /** Whether the controller is initiated asynchronously. */
@@ -51,6 +59,11 @@ export class WebSocketController extends Controller {
             || config.lang;
     }
 
+    /** A reference to the class object. */
+    get Class(): typeof WebSocketController {
+        return <any>this.constructor;
+    }
+
     /** Gets/Sets the DB instance. */
     get db(): DB {
         return this.socket.db;
@@ -68,5 +81,9 @@ export class WebSocketController extends Controller {
     /** Alias of `socket.user` */
     get user(): User {
         return this.socket.user;
+    }
+
+    set user(v: User) {
+        this.socket.user = v;
     }
 }