+++ src/core/middleware/WebSocketSessionHandler.ts
@@ -1,6 +1,6 @@
 import { Server } from "socket.io";
 import { WebSocket } from "../tools/interfaces";
-import { session } from "../bootstrap/session";
+import { session, getHash } from "../bootstrap/session";
 
 export function handleWebSocketSession(io: Server): void {
     io.use((socket, next) => {
@@ -17,6 +17,15 @@ export function handleWebSocketSession(io: Server): void {
             has: (session, key) => key in session,
             deleteProperty: (session, key) => delete session[key]
         });
+
+        let hash = getHash(socket.session);
+
+        // save session to store when the socket is closed.
+        socket.on("disconnected", () => {
+            if (hash !== getHash(socket.session))
+                socket.session.save(() => { });
+        });
+
         next();
     });
 }