+++ HISTORY.md
@@ -1,3 +1,18 @@
+# 0.1.13
+
+## Fixes
+
+- Fix the bug that after calling `res.end()`, `res.send()` or `res.redirect()`,
+    the framework will still try to send data.
+
+## New Features
+
+- **DI (Dependency Injection)** support for controllers, now can set URL params
+    as method parameter, and pass `req: Request`, `res: Response`, or 
+    `socket: WebSocket` in arbitrary orders.
+- Add two special methods `Controller.before()` and `Controller.after()` to 
+    handle any operations before and after the actual method is called.
+
 # 0.1.12
 
 ## Fixes
@@ -7,6 +22,7 @@
 # 0.1.11
 
 ## Fixes
+
 - Fix the bug of `@requireAuth` not working with `WebSocketController`s.
 
 ## New Features